feat(saas): fetch and use roles client-side

+++ saas/{{cookiecutter.project_name}}/webapp/src/routes/my/logout/+page.ts
@@ -1,9 +1,14 @@
+import { browser } from '$app/environment';
 import { currentUser, pb } from '$lib/pocketbase';
+import { rolesStore } from '$lib/roles';
 import { redirect } from '@sveltejs/kit';
 
 export const load = async () => {
-	localStorage.clear();
-	pb.authStore.clear();
-	currentUser.set(null);
-	throw redirect(303, '/');
+	if (browser) {
+		localStorage.clear();
+		pb.authStore.clear();
+		currentUser.set(null);
+		rolesStore.set({});
+		throw redirect(303, '/');
+	}
 };
